added lastfm scrobbler, login screen, track queue, redesign desktop UI, some optimizations

+++ client/src/main/java/com/dismu/music/storages/PlayerBackend.java
@@ -31,7 +31,7 @@ public class PlayerBackend {
         return localInstance;
     }
 
-    public double getPosition() {
+    public long getPosition() {
         return player.getPosition();
     }
